% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitAlignerR-package.R
\docType{package}
\name{SplitAlignerR-package}
\alias{SplitAlignerR}
\alias{SplitAlignerR-package}
\title{SplitAlignerR: Split-Based Branch Mapping and Coordinate Auditing for Phylogenomics}
\description{
An R implementation, currently a scaffold, of the SplitAligner branch-identity coordinate system for phylogenomics under missing taxa and gene-tree discordance. This release ships the deterministic Catnip10 10-tip coordinate-audit benchmark expected oracle outputs as bundled data and a placeholder function so the package installs and checks cleanly. The split-based branch-mapping engine will be added in later versions. See the SplitAligner preprint (Wu 2026, \doi{10.64898/2026.02.24.707838}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wujiaqi06/SplitAlignerR}
  \item Report bugs at \url{https://github.com/wujiaqi06/SplitAlignerR/issues}
}

}
\author{
\strong{Maintainer}: Jiaqi Wu \email{wujiaqi@hiroshima-u.ac.jp}

}
\keyword{internal}
